requirements.txt
dvc~=3.39.0
loguru~=0.7.2
openpyxl~=3.1.2
pandas~=2.0.2
PyYAML~=6.0
seaborn~=0.12.2